Hacks to make changing modes (inline, side-by-side) work
The problem is probably with react's algorithm for DOM manipulation.
The hack for now is to just completely replace the view (i.e. remove
view from DOM) with a new view. This suffers in terms of performance,
but works for now.

Another problem is `componentWillMount()` doesn't always seem to be
called on CheckboxesView so changing to a `componentDidMount()` for
now, which seems to work so far...